+++ src/engine/engine_io.c
@@ -1958,6 +1958,7 @@ const char* mj_validateReferences(const mjModel* m) {
 
     case mjEQ_FLEX:
     case mjEQ_FLEXVERT:
+    case mjEQ_FLEXSTRAIN:
       if (obj1id >= m->nflex || obj1id < 0) {
         return "Invalid model: eq_obj1id out of bounds.";
       }