+++ src/user/user_flexcomp.h
@@ -83,7 +83,7 @@ class mjCFlexcomp {
   double origin[3];               // origin for generating a 3D mesh from a convex 2D mesh
   double mass;                    // total mass of auto-generated bodies
   double inertiabox;              // size of inertia box for each body
-  int equality;                   // create equality constraint, 0:none, 1:edge, 2:vert
+  int equality;                   // create equality constraint, 0:none, 1:edge, 2:vert, 3:strain
   std::string file;               // mesh/gmsh file name
   mjtDof doftype;                 // dof type, all vertices or trilinear interpolation